// src/app/shared/model/LTKM/LtkmCustEmrgncCntctObj.Model.ts
export class LtkmCustEmrgncCntctObj {
    LtkmCustEmrgncCntctId: number;
    LtkmCustId: number;
    ContactPersonName: string;
    MrIdTypeCode: string;
    IdNo: string;
    IdExpiredDt: Date;
    BirthPlace: string;
    BirthDt: Date;
    MotherMaidenName: string;
    MrGenderCode: string;
    MrReligionCode: string;
    MrEducationCode: string;
    MrJobProfessionCode: string;
    MrMaritalStatCode: string;
    MrNationalityCode: string;
    NationalityCountryCode: string;
    MrCustRelationshipCode: string;
    MobilePhnNo1: string;
    MobilePhnNo2: string;
    Email: string;
    Addr: string;
    AreaCode1: string;
    AreaCode2: string;
    AreaCode3: string;
    AreaCode4: string;
    City: string;
    Zipcode: string;
    SubZipcode: string;
    PhnArea1: string;
    Phn1: string;
    PhnExt1: string;
    PhnArea2: string;
    Phn2: string;
    PhnExt2: string;
    PhnArea3: string;
    Phn3: string;
    PhnExt3: string;
    RowVersion: string[];

    constructor() {
        this.LtkmCustEmrgncCntctId = 0,
            this.LtkmCustId = 0,
            this.ContactPersonName = "",
            this.MrIdTypeCode = "",
            this.IdNo = "",
            this.IdExpiredDt = new Date(),
            this.BirthPlace = "",
            this.BirthDt = new Date(),
            this.MotherMaidenName = "",
            this.MrGenderCode = "",
            this.MrReligionCode = "",
            this.MrEducationCode = "",
            this.MrJobProfessionCode = "",
            this.MrMaritalStatCode = "",
            this.MrNationalityCode = "",
            this.NationalityCountryCode = "",
            this.MrCustRelationshipCode = "",
            this.MobilePhnNo1 = "",
            this.MobilePhnNo2 = "",
            this.Email = "",
            this.Addr = "",
            this.AreaCode1 = "",
            this.AreaCode2 = "",
            this.AreaCode3 = "",
            this.AreaCode4 = "",
            this.City = "",
            this.Zipcode = "",
            this.SubZipcode = "",
            this.PhnArea1 = "",
            this.Phn1 = "",
            this.PhnExt1 = "",
            this.PhnArea2 = "",
            this.Phn2 = "",
            this.PhnExt2 = "",
            this.PhnArea3 = "",
            this.Phn3 = "",
            this.PhnExt3 = "",
            this.RowVersion = [];
    }
}